+++ site/vite.config.ts
@@ -1,10 +1,16 @@
 <<<<<<< HEAD
+<<<<<<< HEAD
 export default {};
 =======
+=======
+>>>>>>> origin/w5t1y7-codex/create-top-level-repo-layout
 import { defineConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
 
 export default defineConfig({
   plugins: [vue()]
 });
+<<<<<<< HEAD
 >>>>>>> origin/pybde0-codex/create-top-level-repo-layout
+=======
+>>>>>>> origin/w5t1y7-codex/create-top-level-repo-layout